refactor: User 도메인 및 DTO 패키지 구조 정리 및 관련 코드 수정
