--- src/main/java/com/palangwi/soup/domain/Gender.java
@@ -1,5 +1,0 @@
-package com.palangwi.soup.domain;
-
-public enum Gender {
-    MALE, FEMALE;
-}
