--- src/main/java/com/palangwi/soup/dto/user/UserAdditionalInfoRequest.java
@@ -1,14 +1,0 @@
-package com.palangwi.soup.dto.user;
-
-import com.palangwi.soup.domain.Gender;
-import com.palangwi.soup.security.Role;
-
-import java.time.LocalDate;
-
-public record UserAdditionalInfoRequest(
-        String email,
-        Role role,
-        Gender gender,
-        LocalDate birthDate
-) {
-}
