+++ src/main/java/com/palangwi/soup/dto/user/UserInitSettingResponseDto.java
@@ -1,7 +1,7 @@
 package com.palangwi.soup.dto.user;
 
-import com.palangwi.soup.domain.User;
-import com.palangwi.soup.domain.Gender;
+import com.palangwi.soup.domain.user.User;
+import com.palangwi.soup.domain.user.Gender;
 import com.palangwi.soup.security.Role;
 
 import java.time.LocalDate;
@@ -10,20 +10,20 @@ public record UserInitSettingResponseDto(
         Long userId,
         String email,
         String nickname,
-        String profileImageUrl,
         Role role,
         Gender gender,
-        LocalDate birthDate
+        LocalDate birthDate,
+        String profileImageUrl
 ) {
     public static UserInitSettingResponseDto of(User user) {
         return new UserInitSettingResponseDto(
                 user.getId(),
+                user.getEmail(),
                 user.getNickname(),
-                user.getProfileImageUrl(),
                 user.getRole(),
                 user.getGender(),
                 user.getBirthDate(),
-                user.getEmail()
+                user.getProfileImageUrl()
         );
     }
 }
