+++ src/main/java/com/palangwi/soup/dto/user/UserResponseDto.java
@@ -1,13 +1,22 @@
 package com.palangwi.soup.dto.user;
 
-import com.palangwi.soup.domain.User;
+import com.palangwi.soup.domain.user.Gender;
+import com.palangwi.soup.domain.user.User;
+import com.palangwi.soup.security.Role;
 
-public record UserResponseDto(String nickname,
-                              String profileImageUrl) {
+import java.time.LocalDate;
+
+public record UserResponseDto(String email, String username, String nickname, Role role,
+                              Gender gender, LocalDate birthDate, String profileImageUrl) {
 
     public static UserResponseDto of(User user) {
         return new UserResponseDto(
+                user.getEmail(),
+                user.getUsername(),
                 user.getNickname(),
+                user.getRole(),
+                user.getGender(),
+                user.getBirthDate(),
                 user.getProfileImageUrl()
         );
     }
