+++ src/main/java/com/palangwi/soup/exception/user/UserNotFoundException.java
@@ -1,8 +1,9 @@
 package com.palangwi.soup.exception.user;
 
+import com.palangwi.soup.exception.BaseCustomException;
 import org.springframework.http.HttpStatus;
 
-public class UserNotFoundException extends RuntimeException {
+public class UserNotFoundException extends BaseCustomException {
   @Override
   public String getMessage() {
     return UserExceptionMessage.USER_NOT_FOUND.getMessage();
