+++ src/main/java/com/palangwi/soup/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.palangwi.soup.repository;
 
-import com.palangwi.soup.domain.User;
+import com.palangwi.soup.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,5 +9,6 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
     boolean existsByUsername(String username);
-    Optional<User> findByUsername(String username);
+    Optional<User> findByUsernameAndIsDeletedFalse(String username);
+    boolean existsByNickname(String nickname);
 }
