+++ src/main/java/com/palangwi/soup/security/OAuth2AuthenticationSuccessHandler.java
@@ -1,7 +1,7 @@
 package com.palangwi.soup.security;
 
 import com.palangwi.soup.dto.user.UserInfo;
-import com.palangwi.soup.domain.User;
+import com.palangwi.soup.domain.user.User;
 import com.palangwi.soup.security.Jwt.Claims;
 import com.palangwi.soup.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
